requirements.txt
discord.py==2.6.4
mysql-connector-python==9.5.0
